# openapi.yaml
openapi: 3.0.0
info:
  title: Find My Artisan API
  version: 1.1.0
  description: >
    Access to artisan profiles. Meditrac (Team 2) uses the two GET
    endpoints to find and vet artisans, and POST /artisans/{id}/reviews to
    rate them after a job. Meditrac contacts artisans directly for
    maintenance work at their pharmacies — the booking itself is a phone
    call using the contact details this API exposes, not an API call.
    Meditrac never creates, edits, or holds a booking record on our side.
    The three availability endpoints are used only by our own artisans.

servers:
  - url: http://localhost:5001/api
    description: Local development server
  - url: https://api.findmyartisan.app/api

paths:
  /artisans:
    get:
      summary: List artisans, optionally filtered by trade, county, and same-day availability
      parameters:
        - name: trade
          in: query
          required: false
          description: Filter to artisans practicing this trade.
          schema:
            type: string
            example: plumbing
        - name: county
          in: query
          required: false
          description: Filter to artisans based in this county.
          schema:
            type: string
            example: Nairobi
        - name: availableToday
          in: query
          required: false
          description: Filter to artisans who are free right now (no availability block covers the current time). Must be true or false.
          schema:
            type: boolean
            example: true
      responses:
        '200':
          description: A list of artisans matching the filters (empty list if none match).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ArtisanSummary'
              example:
                - id: 1
                  name: "Wanjiru Kamau"
                  trade: "electrical"
                  county: "Nairobi"
                  verified: true
                  availableToday: true
                - id: 3
                  name: "Achieng Nyambura"
                  trade: "plumbing"
                  county: "Nairobi"
                  verified: true
                  availableToday: false
        '400':
          description: A query parameter has an invalid value.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "invalid_query"
                message: "county must be a known Kenyan county name."

  /artisans/{id}:
    get:
      summary: Get one artisan's full profile, including verification details and contact phone
      parameters:
        - name: id
          in: path
          required: true
          description: The artisan's unique identifier.
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: The artisan's full profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artisan'
              example:
                id: 1
                name: "Wanjiru Kamau"
                trade: "electrical"
                county: "Nairobi"
                verified: true
                availableToday: true
                phone: "+254712000001"
                hourlyRateKes: 900
                verification:
                  issuingBody: "National Industrial Training Authority"
                  certificateId: "NITA-EL-2024-00001"
                  expiresOn: "2027-03-01"
        '404':
          description: No artisan exists with this id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "not_found"
                message: "No artisan with id 1."

  /artisans/{id}/reviews:
    post:
      summary: Leave a review of an artisan after a job
      description: >
        Called by Meditrac after a job agreed by phone is finished. The review
        appears on the artisan's public profile and helps later vetting. This
        records feedback only; it does not create or change a booking.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 3
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewCreate'
            example:
              author: "Meditrac — Westlands branch"
              rating: 5
              comment: "Fixed the leaking sink the same afternoon."
      responses:
        '201':
          description: The review was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
              example:
                id: 58
                artisanId: 3
                author: "Meditrac — Westlands branch"
                rating: 5
                comment: "Fixed the leaking sink the same afternoon."
                createdAt: "2026-09-22T10:30:00Z"
        '400':
          description: A required field is missing, or a value has the wrong type or is out of range.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "invalid_body"
                message: "rating must be a whole number from 1 to 5."
        '404':
          description: No artisan exists with this id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "not_found"
                message: "No artisan with id 999."

  /artisans/{id}/availability:
    post:
      summary: Mark the artisan unavailable for a time window, once a job is agreed by phone
      description: >
        Artisan-facing only — not part of Meditrac's contract. Called by the
        artisan's own account right after agreeing a job over the phone.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvailabilityBlockCreate'
            example:
              start: "2026-09-14T09:00:00Z"
              end: "2026-09-14T12:00:00Z"
      responses:
        '201':
          description: The block was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailabilityBlock'
              example:
                id: 1
                artisanId: 1
                start: "2026-09-14T09:00:00Z"
                end: "2026-09-14T12:00:00Z"
                createdAt: "2026-09-07T10:15:00Z"
        '400':
          description: end is not after start, or the window overlaps an existing block.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "invalid_body"
                message: "end must be after start."
        '404':
          description: No artisan exists with this id.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "not_found"
                message: "No artisan with id 999."

  /artisans/{id}/availability/{blockId}:
    patch:
      summary: Change the time window of an existing unavailability block
      description: >
        Artisan-facing only — not part of Meditrac's contract. For when an
        agreed job's time shifts, not for when it falls through entirely
        (use DELETE for that).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
        - name: blockId
          in: path
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        description: Only the fields being changed need to be sent.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvailabilityBlockUpdate'
            example:
              end: "2026-09-14T13:00:00Z"
      responses:
        '200':
          description: The updated block.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailabilityBlock'
              example:
                id: 1
                artisanId: 1
                start: "2026-09-14T09:00:00Z"
                end: "2026-09-14T13:00:00Z"
                createdAt: "2026-09-07T10:15:00Z"
        '400':
          description: No fields sent, a value is not a date-time, end is not after start, or the new window overlaps another block.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "invalid_body"
                message: "end must be after start."
        '404':
          description: No block exists with this id for this artisan.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "not_found"
                message: "No availability block with id 1."

    delete:
      summary: Remove an unavailability block entirely — the job it was for isn't happening
      description: >
        Artisan-facing only — not part of Meditrac's contract. Different
        from PATCH: PATCH says the job's time shifted, DELETE says the job
        fell through and the window is available again.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            example: 1
        - name: blockId
          in: path
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '204':
          description: The block was removed. No content is returned.
        '404':
          description: No block exists with this id for this artisan.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: "not_found"
                message: "No availability block with id 1."

components:
  schemas:
    ArtisanSummary:
      type: object
      required: [id, name, trade, county, verified, availableToday]
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
        trade:
          type: string
          example: plumbing
        county:
          type: string
          example: Nairobi
        verified:
          type: boolean
        availableToday:
          type: boolean
          description: Computed — true unless the artisan has an availability block (see AvailabilityBlock) covering the current time. Not stored as a raw flag.

    Artisan:
      allOf:
        - $ref: '#/components/schemas/ArtisanSummary'
        - type: object
          required: [phone]
          properties:
            phone:
              type: string
              description: Contact number. Only exposed on the full profile, not the list, so the consumer views an artisan before reaching out.
              example: "+254712000001"
            hourlyRateKes:
              type: number
              example: 800
            verification:
              type: object
              description: Present only when verified is true.
              properties:
                issuingBody:
                  type: string
                  example: "National Industrial Training Authority"
                certificateId:
                  type: string
                  example: "NITA-EL-2024-00001"
                expiresOn:
                  type: string
                  format: date
                  example: "2027-03-01"

    ReviewCreate:
      type: object
      required: [author, rating]
      properties:
        author:
          type: string
          minLength: 1
          maxLength: 100
          description: Who is leaving the review, e.g. the pharmacy branch.
          example: "Meditrac — Westlands branch"
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        comment:
          type: string
          nullable: true
          maxLength: 1000
          example: "Fixed the leaking sink the same afternoon."

    Review:
      type: object
      required: [id, artisanId, author, rating, comment, createdAt]
      properties:
        id:
          type: integer
          example: 58
        artisanId:
          type: integer
          example: 3
        author:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time

    AvailabilityBlockCreate:
      type: object
      required: [start, end]
      properties:
        start:
          type: string
          format: date-time
          example: "2026-09-14T09:00:00Z"
        end:
          type: string
          format: date-time
          example: "2026-09-14T12:00:00Z"

    AvailabilityBlockUpdate:
      type: object
      description: At least one of the two fields must be present.
      properties:
        start:
          type: string
          format: date-time
          example: "2026-09-14T09:00:00Z"
        end:
          type: string
          format: date-time
          example: "2026-09-14T13:00:00Z"

    AvailabilityBlock:
      type: object
      required: [id, artisanId, start, end, createdAt]
      properties:
        id:
          type: integer
          example: 1
        artisanId:
          type: integer
          example: 1
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time

    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: "not_found"
        message:
          type: string
          example: "No artisan with id 1."